Reorganize functions, clean up code, add/update documentation
